# perf updates for __init__(), step(), and log()