Data point - MacBook Pro M4 Max
- sum(__init()__) went from 54.8 seconds to 28.7 seconds
- sum(step()) went from 97.7 seconds to 17.7 seconds
- sum(log()) went from 40.9 seconds to 28.4 seconds

Total, not including pre-__init()__, 197.7 seconds (3:18) down to 76.8 seconds (1:17) - 2.5x

-----
Second data point - Lenovo ThinkPad X31 - Intel Core i7
- sum(__init()__) went from 151.2 to 76.5 seconds
- sum(step()) went from 408.4 to 83 seconds
- sum(log()) went from 64.8 to 104(?) seconds

Total - not including pre-__init__(), 635 seconds (10:35) down to 270 seconds (4:30) - 2.3x